Conectare baza de date

+++ src/main/java/org/example/repository/RepositoryAnimal.java
@@ -4,6 +4,6 @@ import org.example.model.Animal;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface RepositoryAnimal extends JpaRepository<Animal,Integer> {
+public interface RepositoryAnimal extends JpaRepository<Animal,Long> {
 
 }
